Finalizing refactor; prep for multiple notebooks
