--- src/routes/+page.js
@@ -1,6 +1,0 @@
-import { populateSphere } from '../lib/utils';
-
-export function load() {
-  console.log('load');
-  populateSphere();
-}
