--- src/routes/layout.js
@@ -1,3 +1,0 @@
-export const csr = true
-export const ssr = false
-export const prerender = true
